+++ crates/node/src/pruning.rs
@@ -540,6 +540,7 @@ mod tests {
                     .commit_genesis_block(
                         &block,
                         &ChainConfig {
+                            fee_accounting_activation_height: None,
                             chain_id: 1337,
                             genesis_hash: block.hash(),
                         },
@@ -759,6 +760,7 @@ mod tests {
                     .commit_genesis_block(
                         &block,
                         &ChainConfig {
+                            fee_accounting_activation_height: None,
                             chain_id: 1337,
                             genesis_hash: parent,
                         },